+++ .github/bot/src/nightly-version.ts
@@ -1,0 +1,3 @@
+// npx ts-node scripts/bot/src/nightly-version.ts $version
+
+const version = process.argv[2];
