--- scripts/bot/package.json
@@ -1,16 +1,0 @@
-{
-    "name": "swc-bot",
-    "version": "1.0.0",
-    "main": "index.js",
-    "license": "MIT",
-    "private": true,
-    "devDependencies": {
-        "@types/node": "^17.0.17",
-        "ts-node": "^10.5.0",
-        "typescript": "^4.5.5"
-    },
-    "dependencies": {
-        "@octokit/rest": "^18.12.0",
-        "yaml": "^1.10.2"
-    }
-}
